🚚 Reorganize queries
