+++ frontend/app/lib/queries/index.ts
@@ -1,13 +1,9 @@
 export * from "./compose-stacks";
 export * from "./deployments";
-export * from "./environments";
-export * from "./git-apps";
 export * from "./misc";
 export * from "./projects";
-export * from "./server";
 export * from "./server-admin";
 export * from "./services";
 export * from "./shared";
-export * from "./templates";
 export * from "./user";
 export * from "./workspaces";
